Experiment with NSJSONSerialization's stream methods
It looks like they don’t wait for a complete JSON object to be available
on the stream. If I’m wrong, let’s use them.

For me, they only work some of the time (when the whole offer has been
written to the stream before it tries to read)

(If they did wait, they'd be blocking — which might be a little weird.)