Don't show clips in the status menu
Notification center is a better home for them! It prunes old entries,
and displays a configurable number of things, which is exactly what we
were starting to build.

Instead, add some code to discard clips in Core Data whose notifications
are no longer in Notification Center.

I'm a fan of this whole idea, but open to contraindications.